refactor(DocumentRender): バリデーションに関わるクラスを別ファイルに分離
