+++ features/document_render.py
@@ -47,7 +47,6 @@ with open('template.txt', 'rb') as f:
 ```
 """
 
-import re
 from datetime import datetime
 from enum import IntEnum
 from io import BytesIO
@@ -56,7 +55,6 @@ from typing import Annotated, Any, ClassVar, Dict, Optional, Self, Union
 import jinja2
 from jinja2 import Environment, nodes
 from jinja2.runtime import StrictUndefined, Undefined
-from markupsafe import Markup
 from pydantic import (
     AfterValidator,
     BaseModel,
@@ -67,6 +65,8 @@ from pydantic import (
     field_validator,
     model_validator,
 )
+
+from features.validate_template import TemplateSecurityValidator, ValidationState  # type: ignore
 
 
 class ValidationResult(BaseModel):
@@ -135,20 +135,6 @@ class FormatType(IntEnum):
     COMPACT = 4
 
 
-class TemplateValidationError(Exception):
-    """テンプレート検証時に発生するエラー。
-
-    テンプレートの構文やセキュリティ上の問題を示します。
-    以下の場合に発生します：
-    - 禁止されたタグや属性の使用
-    - 再帰的な構造の検出
-    - メモリ使用量の超過
-    - UTF-8デコードエラー
-    """
-
-    pass
-
-
 class TemplateContent(BaseModel):
     """テンプレートの内容を表すモデル。
 
@@ -191,232 +177,6 @@ class TemplateContent(BaseModel):
         if len(self.content.encode("utf-8")) != self.file_size:
             raise ValueError("Content size does not match file size")
         return self
-
-
-class ValidationState:
-    """テンプレートの検証状態を表すクラス。"""
-
-    def __init__(self) -> None:
-        """初期化。"""
-        self._is_valid: bool = True
-        self._error_message: Optional[str] = None
-        self._content: Optional[str] = None
-
-    @property
-    def is_valid(self) -> bool:
-        """検証が成功したかどうかを返す。"""
-        return self._is_valid
-
-    @property
-    def error_message(self) -> Optional[str]:
-        """エラーメッセージを返す。
-
-        Returns:
-            Optional[str]: エラーメッセージ（エラーがない場合はNone）
-        """
-        return self._error_message if self._error_message else None
-
-    @property
-    def content(self) -> Optional[str]:
-        """テンプレート内容を返す。"""
-        return self._content
-
-    def set_error(self, message: Optional[str]) -> None:
-        """エラーメッセージを設定する。
-
-        Args:
-            message: エラーメッセージ（Noneの場合は空文字列に変換）
-        """
-        self._is_valid = False
-        self._error_message = str(message) if message is not None else ""
-
-    def set_content(self, content: str) -> None:
-        """テンプレート内容を設定する。
-
-        Args:
-            content: テンプレート内容
-        """
-        self._content = content
-
-    def reset(self) -> None:
-        """状態をリセットする。"""
-        self._is_valid = True
-        self._error_message = None
-        self._content = None
-
-
-class TemplateSecurityValidator:
-    """テンプレートのセキュリティ検証を行うクラス。
-
-    テンプレートの構文、セキュリティ、および構造を検証します。
-    検証は以下の2段階で実行されます：
-    1. 静的解析（初期検証）
-       - ファイルサイズの検証
-       - エンコーディングの検証
-       - 禁止タグのチェック
-       - 禁止属性のチェック
-       - リテラル値のループ範囲チェック
-    2. ランタイム検証
-       - 再帰的構造の検出
-       - ゼロ除算の検証
-       - 動的なループ範囲の検証
-
-    Attributes:
-        MAX_RANGE_SIZE (int): rangeで許容される最大サイズ（100,000）
-        RESTRICTED_TAGS (Set[str]): 禁止されたタグのセット
-        RESTRICTED_ATTRIBUTES (Set[str]): 禁止された属性のセット
-    """
-
-    # 許可されない属性のリスト（静的検証）
-    RESTRICTED_ATTRIBUTES: ClassVar[set[str]] = {
-        "request",
-        "config",
-        "os",
-        "sys",
-        "builtins",
-        "eval",
-        "exec",
-        "getattr",
-        "setattr",
-        "delattr",
-        "globals",
-        "locals",
-        "__class__",
-        "__base__",
-        "__subclasses__",
-        "__mro__",
-    }
-
-    # 許可されないタグのリスト（静的検証）
-    RESTRICTED_TAGS: ClassVar[set[str]] = {"macro", "include", "import", "extends"}
-
-    # 最大range値
-    MAX_RANGE_SIZE: ClassVar[int] = 100000
-
-    def validate_template(self, ast: nodes.Template) -> ValidationState:
-        """テンプレートのセキュリティを検証する。
-
-        Args:
-            ast: 検証対象のテンプレートAST
-
-        Returns:
-            ValidationState: 検証結果
-        """
-        validation_state = ValidationState()
-
-        # 1. 禁止タグの検証
-        if not self._validate_restricted_tags(ast, validation_state):
-            return validation_state
-
-        # 2. 属性アクセスの検証
-        if not self._validate_restricted_attributes(ast, validation_state):
-            return validation_state
-
-        # 3. ループ範囲の検証（リテラル値のみ）
-        if not self._validate_loop_range(ast, validation_state):
-            return validation_state
-
-        return validation_state
-
-    def _validate_restricted_tags(self, ast: nodes.Template, validation_state: ValidationState) -> bool:
-        """禁止タグを検証する。
-
-        Args:
-            ast: 検証対象のAST
-            validation_state: 検証状態
-
-        Returns:
-            bool: 検証が成功したかどうか
-        """
-        for node in ast.find_all((nodes.Macro, nodes.Include, nodes.Import, nodes.Extends)):
-            tag_name = node.__class__.__name__.lower()
-            if tag_name in self.RESTRICTED_TAGS:
-                validation_state.set_error(f"Template security error: '{tag_name}' tag is not allowed")
-                return False
-        return True
-
-    def _validate_restricted_attributes(self, ast: nodes.Template, validation_state: ValidationState) -> bool:
-        """禁止属性を検証する。
-
-        Args:
-            ast: 検証対象のAST
-            validation_state: 検証状態
-
-        Returns:
-            bool: 検証が成功したかどうか
-        """
-        for node in ast.find_all(nodes.Getattr):
-            if isinstance(node.node, nodes.Name):
-                attr_name = node.node.name
-                if attr_name in self.RESTRICTED_ATTRIBUTES:
-                    validation_state.set_error(f"Template security error: access to '{attr_name}' is restricted")
-                    return False
-        return True
-
-    def _validate_loop_range(self, ast: nodes.Template, validation_state: ValidationState) -> bool:
-        """ループ範囲を検証する。
-
-        Args:
-            ast: 検証対象のAST
-            validation_state: 検証状態
-
-        Returns:
-            bool: 検証が成功したかどうか
-        """
-        for node in ast.find_all(nodes.For):
-            if isinstance(node.iter, nodes.Call) and isinstance(node.iter.node, nodes.Name) and node.iter.node.name == "range":
-                try:
-                    args = node.iter.args
-                    if len(args) == 1:  # range(stop)
-                        stop = self._evaluate_literal(args[0])
-                        if stop > self.MAX_RANGE_SIZE:
-                            validation_state.set_error(
-                                f"Template security error: loop range exceeds maximum limit of {self.MAX_RANGE_SIZE}"
-                            )
-                            return False
-                    elif len(args) >= 2:  # range(start, stop[, step])
-                        start = self._evaluate_literal(args[0])
-                        stop = self._evaluate_literal(args[1])
-
-                        if len(args) > 2:  # step is provided
-                            step = self._evaluate_literal(args[2])
-                            if step == 0:
-                                validation_state.set_error("Template security error: loop step cannot be zero")
-                                return False
-                        else:
-                            step = 1
-
-                        # Calculate the number of iterations
-                        if step > 0:
-                            iterations = (stop - start + step - 1) // step
-                        else:
-                            iterations = (start - stop - step - 1) // -step
-                        if iterations > self.MAX_RANGE_SIZE:
-                            validation_state.set_error(
-                                f"Template security error: loop range exceeds maximum limit of {self.MAX_RANGE_SIZE}"
-                            )
-                            return False
-                except (TypeError, ValueError):
-                    # 動的な値の場合は、ランタイムでの検証に委ねる
-                    continue
-        return True
-
-    def _evaluate_literal(self, node: nodes.Node) -> int:
-        """リテラル値を評価する。
-
-        Args:
-            node: 評価対象のノード
-
-        Returns:
-            int: 評価結果
-
-        Raises:
-            TypeError: リテラル値でない場合
-            ValueError: 評価できない場合
-        """
-        if isinstance(node, nodes.Const) and isinstance(node.value, (int, float)):
-            return int(node.value)
-        raise TypeError("Not a literal value")
 
 
 class ContextValidator(BaseModel):
@@ -748,34 +508,15 @@ class DocumentRender:
             for node in ast.find_all((nodes.Filter,)):
                 if node.name in {"safe", "html_safe"}:
                     # safeフィルターが使用されている場合、内容を検証
-                    if not self._validate_safe_content(node):
-                        self._validation_state.set_error("Template security error: unsafe HTML content detected")
+                    result = self._security_validator.validate_safe_content(node)
+                    if not result.is_valid:
+                        self._validation_state.set_error(result.error_message)
                         return False
         except Exception as e:
             self._validation_state.set_error(f"Template security error: {e!s}")
             return False
 
         return True
-
-    def _validate_safe_content(self, node: nodes.Filter) -> bool:
-        """safeフィルターで使用されるコンテンツの安全性を検証する。
-
-        Args:
-            node: フィルターノード
-
-        Returns:
-            bool: コンテンツが安全な場合はTrue
-        """
-        try:
-            if isinstance(node.node, nodes.Const):
-                # 定数の場合、内容を直接検証
-                content = str(node.node.value)
-                self._html_safe_filter(content)  # 安全性チェックを実行
-            return True
-        except ValueError:
-            return False
-        except Exception:
-            return False
 
     def apply_context(self, context: Dict[str, Any], format_type: int, is_strict_undefined: bool = True) -> bool:
         """テンプレートにコンテキストを適用する。
@@ -809,7 +550,9 @@ class DocumentRender:
             self._is_strict_undefined = is_strict_undefined
 
             # ランタイム検証の実行（再帰的構造の検出を含む）
-            if not self._runtime_validations(self._ast, context):
+            security_result = self._security_validator.validate_runtime_security(self._ast, context)
+            if not security_result.is_valid:
+                self._validation_state.set_error(security_result.error_message)
                 return False
 
             # テンプレートのレンダリング
@@ -843,247 +586,6 @@ class DocumentRender:
             return False
         except Exception as e:
             self._validation_state.set_error(f"Template rendering error: {e!s}")
-            return False
-
-    def _runtime_validations(self, ast: Union[nodes.Template, jinja2.Template], context: Dict[str, Any]) -> bool:
-        """ランタイム検証を実行する。
-
-        Args:
-            ast: 検証対象のテンプレートAST
-            context: テンプレートに適用するコンテキスト
-
-        Returns:
-            bool: 検証が成功したかどうか
-        """
-        try:
-            if ast is None:
-                self._validation_state.set_error("AST is not available")
-                return False
-
-            # ASTの型を確認
-            if isinstance(ast, jinja2.Template) and self._template_content is not None:
-                ast = ast.environment.parse(self._template_content)
-
-            if not isinstance(ast, nodes.Template):
-                self._validation_state.set_error("Invalid AST type")
-                return False
-
-            # 1. 再帰的構造の検出
-            # 変数の割り当てを追跡するための辞書
-            assignments: Dict[str, Any] = {}
-            for node in ast.find_all((nodes.Assign, nodes.For, nodes.Call)):
-                if isinstance(node, nodes.Assign) and isinstance(node.target, nodes.Name):
-                    target_name = node.target.name
-                    try:
-                        # 変数の値を評価
-                        value = self._evaluate_expression(node.node, context, assignments)
-                        assignments[target_name] = value
-                    except Exception as e:
-                        if "recursive structure detected" in str(e):
-                            self._validation_state.set_error("Template security error: recursive structure detected")
-                            return False
-                        continue
-                elif isinstance(node, nodes.For):
-                    # forループ内での再帰的構造の検出
-                    if isinstance(node.iter, nodes.Name):
-                        iter_name = node.iter.name
-                        if iter_name in assignments:
-                            try:
-                                self._evaluate_expression(node.iter, context, assignments)
-                            except Exception as e:
-                                if "recursive structure detected" in str(e):
-                                    self._validation_state.set_error("Template security error: recursive structure detected")
-                                    return False
-                elif isinstance(node, nodes.Call):
-                    # メソッド呼び出しの検証
-                    try:
-                        self._evaluate_expression(node, context, assignments)
-                    except Exception as e:
-                        if "recursive structure detected" in str(e):
-                            self._validation_state.set_error("Template security error: recursive structure detected")
-                            return False
-                        continue
-
-            # 2. ゼロ除算の検証
-            for node in ast.find_all(nodes.Div):
-                try:
-                    right_value = self._evaluate_expression(node.right, context, assignments)
-                    if right_value == 0:
-                        self._validation_state.set_error("Template security error: division by zero is not allowed")
-                        return False
-                except Exception:
-                    continue
-
-            return True
-
-        except Exception as e:
-            self._validation_state.set_error(f"Runtime validation error: {e!s}")
-            return False
-
-    def _evaluate_expression(self, node: nodes.Node, context: Dict[str, Any], assignments: Dict[str, Any]) -> Any:
-        """式を評価する。
-
-        Args:
-            node: 評価対象のノード
-            context: テンプレートに適用するコンテキスト
-            assignments: 変数の割り当て状態
-
-        Returns:
-            Any: 評価結果
-
-        Raises:
-            ValueError: 再帰的構造が検出された場合
-            TypeError: 式の評価に失敗した場合
-        """
-        try:
-            if isinstance(node, nodes.Name):
-                # 変数名の評価
-                name = node.name
-                if name in assignments:
-                    return assignments[name]
-                elif name in context:
-                    return context[name]
-                elif self._is_strict_undefined:
-                    raise TypeError(f"'{name}' is undefined")
-                return None
-
-            elif isinstance(node, nodes.Const):
-                # 定数の評価
-                return node.value
-
-            elif isinstance(node, nodes.List):
-                # リストの評価
-                result = [self._evaluate_expression(item, context, assignments) for item in node.items]
-                if self._is_recursive_structure(result):
-                    raise ValueError("recursive structure detected")
-                return result
-
-            elif isinstance(node, nodes.Dict):
-                # 辞書の評価 - ノードのキーと値を個別に評価
-                result = {}
-                for pair in node.items:
-                    key = self._evaluate_expression(pair.key, context, assignments)
-                    value = self._evaluate_expression(pair.value, context, assignments)
-                    result[key] = value
-                if self._is_recursive_structure(result):
-                    raise ValueError("recursive structure detected")
-                return result
-
-            elif isinstance(node, nodes.Call):
-                # メソッド呼び出しの評価
-                if isinstance(node.node, nodes.Getattr):
-                    # オブジェクトのメソッド呼び出し
-                    obj = self._evaluate_expression(node.node.node, context, assignments)
-                    method_name = node.node.attr
-
-                    # リストのメソッド呼び出しを検証
-                    if isinstance(obj, list) and method_name in ["append", "extend"]:
-                        args = [self._evaluate_expression(arg, context, assignments) for arg in node.args]
-                        # 再帰的構造のチェック
-                        if any(arg is obj or self._is_recursive_structure(arg) for arg in args):
-                            raise ValueError("recursive structure detected")
-                        # メソッドの実行をシミュレート
-                        new_list = obj.copy()
-                        if method_name == "append":
-                            new_list.append(args[0])
-                        elif method_name == "extend":
-                            new_list.extend(args[0])
-                        if self._is_recursive_structure(new_list):
-                            raise ValueError("recursive structure detected")
-                        return new_list
-
-            elif isinstance(node, nodes.Getattr):
-                # 属性アクセスの評価
-                obj = self._evaluate_expression(node.node, context, assignments)
-                if obj is None:
-                    if self._is_strict_undefined:
-                        raise TypeError(f"Cannot access attribute '{node.attr}' of None")
-                    return None
-                return getattr(obj, node.attr)
-
-            elif isinstance(node, nodes.BinExpr):
-                # 二項演算の評価
-                left = self._evaluate_expression(node.left, context, assignments)
-                right = self._evaluate_expression(node.right, context, assignments)
-                if isinstance(node, nodes.Div) and right == 0:
-                    raise ValueError("division by zero is not allowed")
-                return self._evaluate_binary_operation(node, left, right)
-
-            raise TypeError("Cannot evaluate expression")
-        except Exception as e:
-            if "recursive structure detected" in str(e):
-                raise ValueError("recursive structure detected") from e
-            raise TypeError("Cannot evaluate expression") from e
-
-    def _evaluate_binary_operation(self, node: nodes.BinExpr, left: Any, right: Any) -> Any:
-        """二項演算を評価する。
-
-        Args:
-            node: 二項演算ノード
-            left: 左辺の値
-            right: 右辺の値
-
-        Returns:
-            Any: 評価結果
-
-        Raises:
-            TypeError: サポートされていない演算子の場合
-        """
-        if isinstance(node, nodes.Add):
-            return left + right
-        elif isinstance(node, nodes.Sub):
-            return left - right
-        elif isinstance(node, nodes.Mul):
-            return left * right
-        elif isinstance(node, nodes.Div):
-            return left / right
-        elif isinstance(node, nodes.FloorDiv):
-            return left // right
-        elif isinstance(node, nodes.Mod):
-            return left % right
-        elif isinstance(node, nodes.Pow):
-            return left**right
-        raise TypeError(f"Unsupported binary operator: {type(node).__name__}")
-
-    def _is_recursive_structure(self, value: Any) -> bool:
-        """値が再帰的構造かどうかを判定する。
-
-        Args:
-            value: チェック対象の値
-
-        Returns:
-            bool: 再帰的構造の場合はTrue
-        """
-        try:
-            # 循環参照を検出するための集合
-            seen = set()
-
-            def check_recursive(v: Any, depth: int = 0) -> bool:
-                if depth > 100:  # 再帰の深さ制限
-                    return True
-
-                # オブジェクトのIDを取得
-                obj_id = id(v)
-                if obj_id in seen:
-                    return True
-
-                # リストや辞書の場合は、その要素も確認
-                if isinstance(v, (list, dict, set)):
-                    seen.add(obj_id)
-                    try:
-                        if isinstance(v, (list, set)):
-                            result = any(check_recursive(item, depth + 1) for item in v)
-                        else:  # dict
-                            result = any(check_recursive(item, depth + 1) for item in v.values())
-                        if result:
-                            return True
-                    finally:
-                        seen.remove(obj_id)
-
-                return False
-
-            return check_recursive(value)
-        except Exception:
             return False
 
     def _validate_memory_usage(self, content: str) -> bool:
@@ -1149,8 +651,9 @@ class DocumentRender:
             template = env.from_string(template_content)
 
             # ランタイム検証の実行
-            self._runtime_validations(template, context)
-            if not self._validation_state.is_valid:
+            security_result = self._security_validator.validate_runtime_security(template, context)
+            if not security_result.is_valid:
+                self._validation_state.set_error(security_result.error_message)
                 return None
 
             rendered = template.render(**context)
@@ -1275,8 +778,8 @@ class DocumentRender:
 
         # カスタムフィルターの登録
         env.filters["date"] = self._date_filter
-        env.filters["safe"] = self._html_safe_filter  # safeフィルターを安全な実装に変更
-        env.filters["html_safe"] = self._html_safe_filter
+        env.filters["safe"] = self._security_validator.html_safe_filter  # safeフィルターを安全な実装に変更
+        env.filters["html_safe"] = self._security_validator.html_safe_filter
 
         return env
 
@@ -1309,61 +812,3 @@ class DocumentRender:
 
         except (ValueError, TypeError) as e:
             raise ValueError("Invalid date format") from e
-
-    def _html_safe_filter(self, value: str) -> Markup:
-        """HTMLをエスケープせずに出力する。"""
-        return HTMLSafetyValidator.create_safe_markup(value)
-
-
-class HTMLSafetyValidator:
-    """HTMLの安全性を検証するクラス。"""
-
-    # 安全でないHTMLパターン
-    UNSAFE_PATTERNS: ClassVar[list[str]] = [
-        r"<script",  # スクリプトタグ
-        r"javascript:",  # JavaScriptプロトコル
-        r"data:",  # データURIスキーム
-        r"vbscript:",  # VBScriptプロトコル
-        r"on\w+\s*=",  # イベントハンドラ属性
-    ]
-
-    @staticmethod
-    def validate_html_content(value: str) -> bool:
-        """HTMLコンテンツの安全性を検証する。
-
-        Args:
-            value: 検証対象のHTML文字列
-
-        Returns:
-            bool: HTMLが安全な場合はTrue
-
-        Raises:
-            ValueError: 安全でないHTML要素が含まれる場合
-        """
-        if not isinstance(value, str):
-            raise ValueError("HTML content must be a string")
-
-        # 安全でないHTMLパターンをチェック
-        for pattern in HTMLSafetyValidator.UNSAFE_PATTERNS:
-            if re.search(pattern, value, re.IGNORECASE):
-                return False
-        return True
-
-    @staticmethod
-    def create_safe_markup(value: str) -> Markup:
-        """安全なMarkupオブジェクトを作成する。
-
-        Args:
-            value: HTML文字列
-
-        Returns:
-            Markup: 安全なMarkupオブジェクト
-
-        Raises:
-            ValueError: 安全でないHTML要素が含まれる場合
-        """
-        if not HTMLSafetyValidator.validate_html_content(value):
-            raise ValueError("HTML content contains potentially unsafe elements")
-
-        # 安全性が確認されたHTMLのみをMarkupとして返す
-        return Markup("").join(value)
